photo-organizer-backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis & Celery
celery[redis]==5.4.0
redis==5.2.1

# Auth & Security
pyjwt==2.10.1
cryptography==44.0.0
passlib==1.7.4
httpx==0.28.1

# Image Processing
Pillow==11.1.0
imagehash==4.3.2
numpy==2.2.2
python-magic==0.4.27

# AI Providers
anthropic==0.42.0
openai==1.59.7
dashscope==1.20.14

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Storage
boto3==1.36.4

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1
